+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+### 10.41
+ - Fixed error on PMKID hash validation
+ - Created an option to avoid Network Manager force killing on Evil Twin attacks
+ - Code improvements to allow new kind of arguments based plugins
+ - Dsniff Evil Twin captive portal DNS black hole replaced by dnsmasq
+ - Dockerfile rollback to Parrot (Parrot Security image as base) to avoid Arch Glibc + Dockerhub automated builds problem
+
 ### 10.40
  - Removed deprecated route command
  - Fixed error on WPS PINs found in database counter